+++ preferences/keys.js
@@ -136,6 +136,11 @@ export const SettingsKeys = () => {
       widget_type: 'dropdown',
       test: { pointer: 'slide-through', values: [0, 1, 2] },
     },
+    'animation-bounce': {
+      default_value: 0,
+      widget_type: 'scale',
+      test: { pointer: 'slide-through', values: [0, 0.5, 1] },
+    },
     'animation-magnify': {
       default_value: 0,
       widget_type: 'scale',